Split tests for codacy/phpstan happiness
